requirements.txt
marshmallow-dataclass==8.5.3
PSNAWP==2021.9.3
pypresence==4.2.0
pyyaml==5.4.1
query-string==2019.4.13
requests==2.22.0
typepy==1.1.1
urllib3==1.25.3
git+https://github.com/cadewey/winstray#egg=winstray; sys.platform == 'win32'
